requirements.txt
connexion
google-cloud-firestore
gunicorn